% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arber-package.R
\docType{package}
\name{arber-package}
\alias{arber}
\alias{arber-package}
\title{arber: Geospatial Forest and Soil Ecology}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jrarber/arber}
  \item Report bugs at \url{https://github.com/jrarber/arber/issues}
}

}
\author{
\strong{Maintainer}: Jacob Arber \email{jacobrarber@gmail.com} (\href{https://orcid.org/0009-0001-3647-2056}{ORCID})

}
\keyword{internal}
