% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_depth.R
\name{at_depth}
\alias{at_depth}
\title{Properties through soil depth layers}
\usage{
at_depth(var, extent, area = "state", summary = FALSE)
}
\arguments{
\item{var}{The soil variable to download.}

\item{area}{Section of contiguous United States to download.(South, Northeast, Midwest, West)}

\item{summary}{Boolean. If TRUE returns an additional layer as an average of all soil depths.}

\item{Depth}{Discrete soil depth for var. 0_5, 5_15, 15_30, 30_60, 60_100, 100_200}
}
\value{
Multiple layer raster of single soil property.
}
\description{
Raster and summarize soil characteristics through 6 layers of soil depth.
}
\examples{
# Example usage of the function
polaris("clay", "0_5", "~/Downloads/, region = "SE")
polaris("om", "30_60", ~/Users/john/documents/", region = "NW")
}
