% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAD.R
\name{BAD}
\alias{BAD}
\title{Basal Area Discrepency (BAD) values for a given level.}
\usage{
BAD(x, level)
}
\arguments{
\item{x}{The data frame to be used for calculations. Must be at the species level.}

\item{level}{The hierarchical level to calculate BAD values. Often is site, but can vary based on question (genus, etc.).}
}
\value{
A summary with BAD values for each site.
}
\description{
Calculate disproportionate basal areas using Simpson's Diversity. Ranges from 0 (basal area relatively even across level) to 1 (basal area highly disproportionate across level).
}
\examples{
# Example usage of the function
BAD(plotIV, Site)
}
