% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocurs.R
\name{cocurs}
\alias{cocurs}
\title{Co-Occurring Species}
\usage{
cocurs(
  species,
  data = NULL,
  diag = FALSE,
  colors = NULL,
  silent = FALSE,
  method = "square"
)
}
\description{
Calculates what species occur with a given species, and how often.
 Also creates a matrix of how often the co-occurring species occur with eachother.
}
