% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_to_area.R
\name{cut_to_area}
\alias{cut_to_area}
\title{Cutting Rasters to State and Regional Boundaries}
\usage{
cut_to_area(raster, extent, area = "state")
}
\arguments{
\item{raster}{The uncropped raster to be cut.}

\item{extent}{The state or region to crop to. Regions: "South", "Northeast", "West", "Midwest"}
}
\value{
A cropped raster
}
\description{
A more detailed description of what the function does.
}
\examples{
# Example usage of the function
cut(rawraster, "Arkansas", area = "state")
cut(rawraster, "SW", area = "region")
}
