% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_from_raster.R
\name{data_from_raster}
\alias{data_from_raster}
\title{Extract data from a raster and add it to your data frame}
\usage{
data_from_raster(x, raster, variable)
}
\arguments{
\item{x}{The data frame with latitudes and longitudes in columns "LAT" and "LON"}

\item{raster}{A loaded raster with data to be extracted.}

\item{variable}{The desired name of the column containing the contents extracted from the raster.}
}
\value{
The inputted dataframe with an added column of name 'variable' with values from the raster.
}
\description{
Get values from a raster that match latitudes and longitudes in your data frame.
}
\examples{
# Example usage of the function
sites_withsand <- data_from_raster(sitedata, sandraster, "Sand")
}
