% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_test.R
\name{dist_test}
\alias{dist_test}
\title{Test the Distribution of Data}
\usage{
dist_test(data)
}
\arguments{
\item{data}{The data for the distribution. If a dataframe is supplied, only numeric columns will be displayed.}
}
\value{
Provides visual graphs of the data and summary statistical output, including mean and standard deviation.
}
\description{
Visually the distribution of data in a vector or a data frame.
}
