% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoregions.R
\name{ecoregions}
\alias{ecoregions}
\title{US EPA Ecoregion names and codes}
\usage{
ecoregions(level = 3)
}
\arguments{
\item{level}{description}
}
\value{
description
}
\description{
A helper function that provides the codes and names of EPA Ecoregions at a given level. Best used as reference for 'ecoshapes()'.
}
\examples{
# example code

}
