% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polaris.R
\name{polaris}
\alias{polaris}
\title{Downloading POLARIS soil data for a given US region}
\usage{
polaris(
  var,
  depth,
  path,
  region,
  crop = FALSE,
  clear.files = FALSE,
  download = TRUE
)
}
\arguments{
\item{var}{The soil variable to download.}

\item{path}{The folder path for files to be downloaded to.}

\item{region}{Section of contiguous United States to download.("South", "Northeast", "Midwest", "West")}

\item{crop}{boolean. If TRUE raster is returned cropped to the US Census Bureau region.}

\item{clear.files}{Boolean. If TRUE, downloaded files are deleted from path after merged into regional raster.}

\item{download}{Boolean. FALSE only if error occurred after downloading and need to remerge.}

\item{Depth}{Discrete soil depth for var. "0_5", "5_15", "15_30", "30_60", "60_100", "100_200".}
}
\value{
Raster values with 30m resolution.
}
\description{
Returning a raster of soil characteristics of four regions of the United States (NE, SE, WE, MW).
}
\examples{
# Example usage of the function
southeast_clay_0_5 <- polaris("clay", "0_5", "~/Downloads/, region = "SE")
northwest_om_30_60 <- polaris("om", "30_60", ~/Users/john/documents/", region = "NW")
}
\seealso{
Other soils: 
\code{\link{polaris2}()}
}
\concept{soils}
