% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polaris2.R
\name{polaris2}
\alias{polaris2}
\title{Downloading state-specific soil properties from POLARIS}
\usage{
polaris2(var, depth, path, state, download = TRUE)
}
\arguments{
\item{var}{The soil variable to download.}

\item{path}{The folder path for files to be downloaded to.}

\item{state}{State data to be downloaded.}

\item{download}{boolean. Determines if the files need to be downloaded or just merged.}

\item{Depth}{Discrete soil depth for var. "0_5", "5_15", "15_30", "30_60", "60_100", "100_200".}
}
\value{
State-wide raster file of soil data.
}
\description{
Returning a state-specific raster of soil characteristics from POLARIS web repisoitory.
}
\examples{
# Example usage of the function
AR_clay_0_5 <- polaris2("clay", "0_5", "~/Downloads/, state = "Arkansas")
ND_om_30_60 <- polaris2("om", "30_60", ~/Users/john/documents/", state = "North Dakota")
}
\seealso{
Other soils: 
\code{\link{polaris}()}
}
\concept{soils}
