% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psco.R
\name{psco}
\alias{psco}
\title{Probability of Species Co-Occurring}
\usage{
psco(sp1, sp2, data = NULL)
}
\arguments{
\item{sp1}{The first species for co-occurrence.}

\item{sp2}{The second species for co-occurrence.}

\item{data}{The data frame for the co-occurence proability to calculate from. If left blank, data will calculate from USFS FIA data.}
}
\value{
Along with the probability, a list is supplied containing information for both species.
}
\description{
Calculates the probability of two species to be in a given plot. Value based
 on US FIA data, or a supplied data frame.
}
\examples{
# Example usage of the function
tadi.nyac <- pcso("Taxodium_distichum", "Nyssa_aquatica")
}
