% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randnova.R
\name{randnova}
\alias{randnova}
\title{Randomization Analysis of Variance}
\usage{
randnova(
  formula,
  data = NULL,
  iterations = 9999,
  post.hoc = FALSE,
  plot = TRUE,
  all.plots = FALSE,
  alpha = 0.05,
  post.hoc.test = "mean.difference"
)
}
\description{
Randomization Analysis of Variance
}
