% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{shape}
\alias{shape}
\title{Load in shape files for US regions or states as spatVectors}
\usage{
shape(region, as.vector = FALSE, columns = FALSE)
}
\arguments{
\item{region}{The area in the US to output a shapefile for. Can be a region ("West", "South", "Midwest", "Northeast") or a state (including DC).}

\item{as.vector}{boolean. If TRUE, will return a spatVector of the chosen area's shapefile. If FALSE, will just return the shapefile.}

\item{columns}{boolean. If FALSE, will remove all columns from the shapefile except for row names. FALSE by default.}
}
\value{
A loaded shapefile or spatVector for the given area.
}
\description{
Easily make available shapefiles and spatVectors of states and regions in the United States. Any state, and any of five regions ("West", "South", "Northeast", "Midwest") are used to provide built-in shapes. 'columns = TRUE' also provides area of land and water in each state.
}
\examples{
# Example usage of the function
plot(shape("Arkansas"))
LA <- shape("Louisiana", as.vector = TRUE)
South <- shape("South")
}
