% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape2.R
\name{shape2}
\alias{shape2}
\title{Regional shapefiles and spatVectors of the US with interior state borders}
\usage{
shape2(region, style = "states", as.vector = FALSE, columns = FALSE)
}
\arguments{
\item{region}{The area in the US to output a shapefile for. Can be a region ("West", "South", "Midwest", "Northeast")}

\item{style}{Either "states" or "ecoregion" to decide the interior borders. "ecoregions" still returns state boundaries."states" by default.}

\item{as.vector}{boolean. If TRUE, will return a spatVector of the chosen area's shapefile. If FALSE, will just return the shapefile.}

\item{columns}{boolean. TRUE by default, removes all columns from the shapefile except for row names.}
}
\value{
A loaded shapefile or spatVector for the given area.
}
\description{
Quickly retrieve shapefiles of United States regions with interior borders of states with an option to overlay level three EPA ecoregions.
}
\examples{
South_Ecoregion_shapes <- shape2("South", style = "ecoregion")
West_spatvector <- shape2("West", as.vector = TRUE)
}
