% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteFIA.R
\name{siteFIA}
\alias{siteFIA}
\title{Summarize tree surveys into a site-level dataframe}
\usage{
siteFIA(x, sites)
}
\arguments{
\item{x}{The data frame with sites and species to be summarized.}

\item{sites}{The column of the data frame that differentiates sites.}
}
\value{
A site-level summary as a data frame.
}
\description{
Summarizes plot characteristics with species and calculates new importance values (IV)
}
\examples{
# Example usage of the function
sites <- siteFIA(plotIV, ID)
}
\seealso{
Other treeFIA: 
\code{\link{sites_with}()},
\code{\link{treeFIA}()}
}
\concept{treeFIA}
