% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_with.R
\name{sites_with}
\alias{sites_with}
\title{Get site FIA plot information where given species are present}
\usage{
sites_with(present, absent = NULL, data = NULL, exact = FALSE)
}
\arguments{
\item{present}{The species (as "Genus_specific"), or multiple with c(), used to choose the sites.}

\item{absent}{The species (as "Genus_specific"), or multiple with c(), to exclude in site selection.}

\item{exact}{boolean. FALSE by default. If TRUE, will only return sites with all of the listed species present.}

\item{x}{Optional. The data to be filtered from. If left blank, will filter from the entire FIA database.}
}
\value{
Dataframe with species level information at plots where the chosen species are present.
}
\description{
Use USFS FIA style data to get information on sites where a species (or multiple species) are present.
}
\examples{
# Example usage of the function
cypress_sites <- sites_with("Taxodium_distichum")
cypress_tupelo <- sites_with(c("Taxodium_distichum","Nyssa_aquatica"), mydata)
cypress_and_tupelo <- sites_with(c("Taxodium_distichum","Nyssa_aquatica"), mydata, exact = TRUE)
}
\seealso{
Other treeFIA: 
\code{\link{siteFIA}()},
\code{\link{treeFIA}()}
}
\concept{treeFIA}
