% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilUS.R
\name{soilUS}
\alias{soilUS}
\title{Mapping soil properties across the contiguous United States}
\usage{
soilUS(var, depth)
}
\arguments{
\item{var}{The soil variable to download.}

\item{Depth}{Discrete soil depth for var. 0_5, 5_15, 15_30, 30_60, 60_300.}
}
\value{
Raster of variable at chosen depth across the continental US.
}
\description{
Returning a raster of the variable mean at depth across the whole continental United States with POLARIS soil rasters at 30m resolution.
}
\examples{
# Example usage of the function
US_clay_raster <- soilUS("clay", "0_5")
US_om_raster_30_60cm <- soilUS("om", "30_60")
}
