% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sympatmat.R
\name{sympatmat}
\alias{sympatmat}
\title{Sympatry Matrices for Tree Species}
\usage{
sympatmat(species.list, method = c("proportion", "basal"), data = NULL)
}
\arguments{
\item{species.list}{A vector of species to include in the matrix.}

\item{method}{Currently one of 'proportional' (calculated based on proportion of shared plots) or 'basal' (calculated by basal area in shared plots)}

\item{data}{The data to be used, including columns 'Species', 'ID', 'BA'}

\item{summary}{Boolean, False by default. If true will provide a matrix of half the length with summaries of the species-species sympatries out of 2.}
}
\value{
A list of length 1 (or 2 with summary plot) and plots the correlation matrix with 'corrplot'. Matrix is read with by the amount of plots with species n (column) present out of total plots with species m (row).
}
\description{
Create and plot matrices of relative sympatry
}
