% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeFIA.R
\name{treeFIA}
\alias{treeFIA}
\title{Load in tree survey data from USFS Forest Inventory and Analysis database}
\usage{
treeFIA(filter.by = "none", value, native.only = FALSE)
}
\arguments{
\item{filter.by}{How the data should be filtered by ("none", "state", "region"). If left blank, will return the unfiltered FIA database.}

\item{value}{Only if 'filter.by' is chosen. The criteria for which to match the filter.}

\item{native.only}{boolean. FALSE by default. If TRUE, will only return native species.}
}
\value{
Raster values with 30m resolution.
}
\description{
Load in a dataframe of forest survey data in its entirety, or filtered by a specific criteria (currently only state or species).
}
\examples{
# Example usage of the function
treeIV <- treeFIA()
ArkansasTreeIV <- treeFIA(filter.by = "state", "Arkansas")
}
\seealso{
Other treeFIA: 
\code{\link{siteFIA}()},
\code{\link{sites_with}()}
}
\concept{treeFIA}
