% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeheight.R
\name{treeheight}
\alias{treeheight}
\title{Maximum Tree Height Model}
\usage{
treeheight(
  wood.dens.stem,
  wood.dens.root,
  assimilation,
  leaf.mass.area = 60,
  growing.season = 210,
  daylength = 10,
  years = NULL,
  echo = F,
  keep.years = T,
  respir.leaf.m = 0.002136,
  respir.stem.m = 9.23333e-05,
  respir.root.m = 0.001051667,
  respir.leaf.a = 0.2,
  respir.stem.v = 60,
  respir.root.v = 100
)
}
\arguments{
\item{wood.dens.stem}{Stem wood density}

\item{wood.dens.root}{Root wood density}

\item{assimilation}{Leaf-level rate of CO2 assimilation, in umol_m-2_s-1}

\item{leaf.mass.area}{Leaf mass area}

\item{growing.season}{Growing season length in days}

\item{daylength}{Daylength in hours during the growing season}

\item{years}{A specified number of years to let the individual grow. If left blank, will grow until it reaches maximum height.}

\item{keep.years}{boolean. If true will return a dataframe with each years associated growth and respiration broken down by main organs (roots, stems, leaves).}

\item{respir.leaf.m}{Mass-based leaf respiration rate, in umol_g-1_s-1.}

\item{respir.stem.m}{Mass-based stem respiration rate, in umol_g-1_s-1.}

\item{respir.root.m}{Mass-based root respiration rate, in umol_g-1_s-1.}

\item{respir.leaf.a}{Area-based leaf respiration rate, in umol_m-2_s-1.}

\item{respir.stem.v}{Volume-based stem respiration rate, in umol_m-3_s-1.}

\item{respir.root.v}{Volume-based root respiration rate, in umol_m-3_s-1.}
}
\value{
Returns the maximum height, how long it took to grow, and a data frame of each years growth assessment (optional).
}
\description{
Model maximum tree height for an individual given wood density, assimilation, and respiration.
}
